saas-boilerplate/backend/pyproject.toml
[project]
name = "saas-boilerplate"
version = "0.1.0"
description = "Django SaaS Boilerplate"
requires-python = ">=3.11"
dependencies = [
    "Django>=5.2,<6.0",
    "djangorestframework>=3.16",
    "django-cors-headers>=4.3",
    "django-allauth>=0.61",
    "django-filter>=24.1",
    "django-ratelimit>=4.1.0",
    "celery>=5.5,<6.0",
    "redis>=5.0",
    "stripe>=8.0",
    "psycopg2-binary>=2.9",
    "gunicorn>=21.2",
    "uvicorn>=0.27",
    "sentry-sdk>=1.40",
    "dj-database-url>=2.1",
    "drf-nested-routers>=0.93.5",
    "drf-spectacular>=0.27.1",
    "pytest-django>=4.8.0",
    "factory-boy>=3.3.0",
    "pyotp>=2.9.0",
    "qrcode>=7.4.2",
    "PyJWT>=2.8.0",
    "cryptography>=42.0.0",
    "argon2-cffi>=23.1.0",
    "python-json-logger>=2.0.7",
    "django-environ>=0.11.2",
    "whitenoise>=6.7.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.3",
    "mypy>=1.10.0",
    "django-stubs>=5.0.0",
    "django-stubs-ext>=5.0.0",
    "types-requests>=2.32.0.20240622",
    "types-redis>=4.6.0.20240311",
    "pre-commit>=3.8.0",
    "locust>=2.20.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["apps", "config"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["apps", "config"]
extend-exclude = ["migrations"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "DJ"]
ignore = ["DJ001"]  # Allow model Meta.verbose_name defaults
preview = true
per-file-ignores = {"config/settings/*.py" = ["F403", "F405"]}

[tool.ruff.lint.isort]
known-first-party = ["apps", "config"]
combine-as-imports = true

[tool.ruff.format]
indent-style = "space"
quote-style = "single"

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
warn_unused_ignores = true
ignore_missing_imports = true
plugins = ["mypy_django_plugin.main"]

[mypy.plugins.django-stubs]
django_settings_module = "config.settings.development"
